task 1/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
cors==1.0.1
fastapi-cors==0.0.6
numpy==1.24.3
requests==2.31.0